Improve syntax for snippet definition
Make the syntax for snippet definition a little bit more consistent
by making the instructions named symmetrically @@snippet_begin and
@@snippet_end.

The end instruction doesn't require an argument anymore. Blocks
can't be nested anyway, so this makes the instruction a bit shorter
and avoids having to repeat the name of the snippet.